Refactoring scheduling pages.
