+++ src/components/schedulePages/AddSolution.js
@@ -2,13 +2,13 @@ import { Box, InputLabel, FormControl, MenuItem, Select, Chip, OutlinedInput, Te
 //import { AsyncTaskManager } from 'builder-util';
 import { React, useState } from 'react';
 import { FaTimes } from 'react-icons/fa';
-import './../assets/styles/HomePage.css';
-import './../assets/styles/SideNav.css';
-import './../assets/styles/AddPages.css';
-import SideNavigation from './SideNavigation.js';
-import TopBar from './TopBar.js';
+import './../../assets/styles/HomePage.css';
+import './../../assets/styles/SideNav.css';
+import './../../assets/styles/AddPages.css';
+import SideNavigation from '../SideNavigation.js';
+import TopBar from '../TopBar.js';
 
-import * as AlgoFunction from './../services/algorithmServices/mainAlgorithmService';
+import * as AlgoFunction from './../../services/algorithmServices/mainAlgorithmService';
 
 
     /**
